improved UX and chunk requests logic

+++ wms2tiff.py
@@ -1,8 +1,10 @@
 from osgeo import gdal
 from xml.etree import ElementTree as ET
+from datetime import datetime, timedelta
 import requests
 import numpy as np
 import os
+import time
 import warnings
 
 warnings.filterwarnings("ignore")
@@ -12,6 +14,16 @@ PASSWORD = os.getenv('PASSWORD')
 
 gdal.SetConfigOption('GDAL_CACHEMAX', '2048')
 os.environ['GDAL_HTTP_TIMEOUT'] = '600'
+
+n_of_quadrants = 16 # 16, 8, 4, 2
+wh_size = 25000 # 10000, 25000, 50000
+
+delete_temp_files = True
+
+step_begin = None
+step_end = None
+
+start_time = datetime.now()
 
 def get_token():
     url = "https://webgis.abdac.it/portal/sharing/rest/generateToken"
@@ -32,6 +44,10 @@ def get_token():
 
     g_token = response.json()["token"]
     return g_token
+
+def clear_previous_lines(n=2):
+    for _ in range(n):
+        print('\033[F\033[K', end='')
 
 def print_dataset_info(dataset):
     if dataset is None:
@@ -79,11 +95,11 @@ def get_capabilities(base_url, use_token=False):
             
     except Exception as e:
         print(f'Error: {str(e)}')
+        return None
 
     capabilities = ET.fromstring(response.content)
 
     cap_dict = {}
-
     namespace = {'ns0': 'http://www.opengis.net/wms'}
 
     cap_dict['max_width'] = capabilities.find('.//ns0:MaxWidth', namespace).text
@@ -92,24 +108,24 @@ def get_capabilities(base_url, use_token=False):
     time_dimension = capabilities.find('.//ns0:Dimension[@name="time"]', namespace)
     cap_dict['time'] = time_dimension.attrib['default'] if time_dimension is not None else None
 
-    minx = capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('minx')
-    maxx = capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('maxx')
-    miny = capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('miny')
-    maxy = capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('maxy')
+    minx = float(capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('minx'))
+    maxx = float(capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('maxx'))
+    miny = float(capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('miny'))
+    maxy = float(capabilities.find('.//ns0:BoundingBox', namespace).attrib.get('maxy'))
 
-    cenx = str((float(minx) + float(maxx))/2)
-    ceny = str((float(miny) + float(maxy))/2)
+    step_x = (maxx - minx) / n_of_quadrants
+    step_y = (maxy - miny) / n_of_quadrants
 
-    cap_dict['bboxes'] = [
-        # Top-left quadrant (min y, min x, center y, center x)
-        f'{miny},{minx},{ceny},{cenx}',  
-        # Top-right quadrant (min y, center x, center y, max x)
-        f'{miny},{cenx},{ceny},{maxx}',  
-        # Bottom-left quadrant (center y, min x, max y, center x)
-        f'{ceny},{minx},{maxy},{cenx}',  
-        # Bottom-right quadrant (center y, center x, max y, max x)
-        f'{ceny},{cenx},{maxy},{maxx}'  
-    ]
+    cap_dict['bboxes'] = []
+
+    for i in range(n_of_quadrants):
+        for j in range(n_of_quadrants):
+            bbox_minx = minx + j * step_x
+            bbox_maxx = minx + (j + 1) * step_x
+            bbox_miny = miny + i * step_y
+            bbox_maxy = miny + (i + 1) * step_y
+
+            cap_dict['bboxes'].append(f'{bbox_miny},{bbox_minx},{bbox_maxy},{bbox_maxx}')
 
     cap_dict['title'] = capabilities.find('.//ns0:Title', namespace).text
     return cap_dict
@@ -161,7 +177,7 @@ def set_transparency(input_tiff, output_tiff, transparency):
 
 def merge_tiffs(files, output_file):
     try:
-        print("Merging TIFF files...")
+        print("Merging TIFF files...", end='\r', flush=True)
 
         warp_options = gdal.WarpOptions(
             format='GTiff',
@@ -170,16 +186,26 @@ def merge_tiffs(files, output_file):
 
         gdal.Warp(destNameOrDestDS=output_file, srcDSOrSrcDSTab=files, options=warp_options)
 
-        print(f"GeoTIFF files merged successfully into {output_file}.")
+        print(f"TIFF files merged successfully into {output_file}", end='\r', flush=True)
 
-        for file in files:
-            if os.path.exists(file):
-                os.remove(file)
+        if delete_temp_files:
+            for file in files:
+                if os.path.exists(file):
+                    os.remove(file)
+            if os.listdir('temp') == []:
+                os.removedirs('temp')
+
     except Exception as e:
         print(f"Error during merging: {str(e)}")
 
+
 def download_wms_layer(wms_url, output_tiff):
+    global step_begin
+    global step_end
+
     output_files = []
+    time_diffs = []
+
     try:
         capabilities = get_capabilities(wms_url, use_token=True)
 
@@ -189,13 +215,14 @@ def download_wms_layer(wms_url, output_tiff):
         width = max_width
         height = max_height
 
-        width, height = 50000, 50000
+        width, height = wh_size, wh_size
+
+        if 'temp' not in os.listdir():
+            os.mkdir('temp')
+
+        temp_output_tiff = os.path.join('temp', output_tiff)
 
         for idx, bbox in enumerate(capabilities['bboxes']):
-            
-            if capabilities['time']:
-                time = capabilities['time']
-
             options = [
                 '-co', 'ALPHA=YES',
                 '-co', 'TILED=YES',
@@ -213,29 +240,70 @@ def download_wms_layer(wms_url, output_tiff):
                 print("GDAL failed to open the WMS dataset.")
                 return
             
-            print('gdal.Translate...')
+            tot_quadrants = n_of_quadrants * n_of_quadrants
             i = idx + 1
-            gdal.Translate(output_tiff.replace('.tiff', f'_{i}.tiff'), wms_dataset, format='GTiff', width=width, height=height, options=options)
-            output_files.append(output_tiff.replace('.tiff', f'_{i}.tiff'))
-            print('Elaborating tiff file...')
-            set_transparency(output_tiff.replace('.tiff', f'_{i}.tiff'), output_tiff.replace('.tiff', f'_{i}.tiff').replace('.tiff', '_transp.tiff'), transparency=0.5)
-            output_files.remove(output_tiff.replace('.tiff', f'_{i}.tiff'))
-            if output_tiff.replace('.tiff', f'_{i}.tiff') in os.listdir():
-                os.remove(output_tiff.replace('.tiff', f'_{i}.tiff'))
-            output_files.append(output_tiff.replace('.tiff', f'_{i}.tiff').replace('.tiff', '_transp.tiff'))
+
+            step_begin = time.time()
+
+            if time_diffs:
+                avg_time_per_quadrant = sum(time_diffs) / len(time_diffs)
+                remaining_quadrants = tot_quadrants - i + 1
+                eta = avg_time_per_quadrant * remaining_quadrants
+                if eta < 60:
+                    eta_finish = datetime.now() + timedelta(seconds=eta)
+                    eta_str = f'~ {int(eta)} seconds remaining (finishes downloading at around: {eta_finish.strftime("%H:%M:%S")})'
+                elif 60 <= eta < 3600:
+                    eta_finish = datetime.now() + timedelta(minutes=eta/60)
+                    eta_str = f'~ {round(eta/60, 2)} minutes remaining (finishes downloading at around: {eta_finish.strftime("%H:%M:%S")})'
+                elif 3600 <= eta:
+                    eta_finish = datetime.now() + timedelta(hours=eta/3600)
+                    eta_str = f'~ {round(eta/3600, 2)} hours remaining (finishes downloading at around: {eta_finish.strftime("%H:%M:%S")})'
+            else:
+                eta_str = ''
+
+            print('\r' + (' ' * 150), end='', flush=True)
+            print(f'\rDownloading {i}/{tot_quadrants}...  {eta_str}', end='', flush=True)
+
+            output_file_with_idx = temp_output_tiff.replace('.tiff', f'_{i}.tiff')
+            gdal.Translate(output_file_with_idx, wms_dataset, format='GTiff', width=width, height=height, options=options)
+            
+            output_files.append(output_file_with_idx)
+
+            set_transparency(output_file_with_idx, output_file_with_idx.replace('.tiff', '_transp.tiff'), transparency=0.5)
+            output_files.remove(output_file_with_idx)
+            if os.path.exists(output_file_with_idx):
+                os.remove(output_file_with_idx)
+            output_files.append(output_file_with_idx.replace('.tiff', '_transp.tiff'))
+
+            step_end = time.time()
+
+            elapsed = step_end - step_begin
+            time_diffs.append(elapsed)
+
+        print('\r' + ' ' * 150, end='\r', flush=True)
 
         merge_tiffs(output_files, output_tiff)
     except Exception as e:
         print(f"Error: {str(e)}")
 
 if __name__ == "__main__":
+    print('------WMS to TIFF------')
     
     # base_wms_url = 'https://webgis.abdac.it/server/services/ETT_PERIC_FRANA_DISTR/MapServer/WMSServer'
     # base_wms_url = 'https://webgis.abdac.it/server/services/Pericolosit%C3%A0_idraulica_Progetto_PAI_distr___solo_fluviale___bozza_/MapServer/WMSServer'
 
     base_wms_url = input('Please write the URL to the WMS Layer:\n')
-    output_path = input('Please write a name for the output .tiff file:\n')
+
+    clear_previous_lines(n=2)
+
+    output_path = input('Please write a name for the output tiff file:\n')
     if not output_path.endswith('.tiff'):
         output_path += '.tiff'
 
+    clear_previous_lines(n=2)
+    
+    print(f'Job started at {start_time}' + ' ' * 50)
     download_wms_layer(base_wms_url, output_path)
+    print(f'Job finished at {datetime.now()}')
+    cwd = os.getcwd()
+    print(f'File saved as {output_path} in {cwd}')
